// src/app/shared/models/instructor.ts
export interface IInstructor{
    id: 0;
    first_name: string;
    second_name: string;
    last_name: string;
    description: string;

    phone: string;
    experience: number;
    // specialization: string;
    cv_path:string;
    img_path:string;
    is_aproved:string;
email:string;

    // "id": 1,
    // "first_name": "ali",
    // "second_name": "khalifa",
    // "last_name": "elsaeed",
    // "description": "ksanckjbsajk ajksbckjsa askjjbcjkas sa",
    // "experience": 3.5,
    // "phone": "01122895762452",
    // "cv_file": "Ps_instructor-606f06534a128.pdf",
    // "img": "Ps_instructor-606f06534a4e6.jpg",
    // "lecturers_specialty_id": 2,
    // "img_path": "http://127.0.0.1:8000/uploads/Ps_instructors/img/Ps_instructor-606f06534a4e6.jpg",
    // "cv_path": "http://127.0.0.1:8000/uploads/Ps_instructors/c.v/Ps_instructor-606f06534a128.pdf"




    
}